style: cleaning up code
