+++ apps/web/src/components/common/sidebar/sections/projects/create-project-modal.tsx
@@ -107,13 +107,6 @@ function CreateProjectModal({ open, onClose }: CreateProjectModalProps) {
                     pattern="[A-Z0-9]+"
                     required
                   />
-                  <Button
-                    type="button"
-                    onClick={() => setSlug(generateProjectSlug(name))}
-                    className="bg-zinc-100 text-zinc-900 hover:bg-zinc-200 dark:bg-zinc-800 dark:text-zinc-100 dark:hover:bg-zinc-700"
-                  >
-                    Generate
-                  </Button>
                 </div>
                 <p className="mt-1 text-xs text-zinc-500 dark:text-zinc-400">
                   This key will be used for ticket IDs (e.g., ABC-123)
