+++ apps/web/src/types/workspace-user/index.ts
@@ -1,0 +1,9 @@
+type WorkspaceUser = {
+  userEmail: string;
+  userName: string | null;
+  joinedAt: Date;
+  status: string;
+  role: string;
+};
+
+export default WorkspaceUser;
