+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kaneo",
-  "private": true,
+  "private": false,
   "type": "module",
   "scripts": {
     "build": "turbo build",
@@ -23,5 +23,5 @@
     "apps/*",
     "packages/*"
   ],
-  "version": "0.7.0"
+  "version": "0.0.0"
 }
